==> Scripts/1. Data import and descriptive analysis.ipynb <==
# 00. Table of contents
 - Importing libraries
 - Importing Dataset
 - Exploring Dataset with basic commands

# 01. Importing Libraries

# STEP 3 - Import libraries
import pandas as pd
import numpy as np
import os 

# 02. Importing Dataset

# STEP 4 - defining path
path = r'C:\Users\viki\Documents\Data Analytics\Immersion\Achievement 4\Instacart Basket Analysis'

#STEP 5a - defining columns needed
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# STEP 5b -import dataset
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

# 03. Exploring Data

# STEP 6 - printing first 5 rows
df.head()

# STEP 7 - importing products dataset
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'),index_col = False)

# STEP 8a - Print the first 20 rows
df_prods.head(20)

# STEP 8b - Print the last 35 rows
df_prods.tail(35)

# STEP 9 - Print the names of the columns
df_prods.columns

# STEP 10 - Print the number of rows and columns
df_prods.shape

# STEP 10a
df_prods.describe()

# STEP 10b
df_prods.dtypes

# Answers Step 10:
maximum value of the “aisle_id” column: 134 ;
data type of the “department_id” column: integer
