==> Scripts/2. Data Wrangling .ipynb <==
# 00. Table of contents
 - Importing libraries
 - Importing Dataset
 - Data Wrangling do-along of Exercise 4.4
     1. Dropping columns
     2. Renaming columns
     3. Changing Variables' Datatype
     4. Transposing Data
     5. Data Dictionary
     6. Subsetting
     7. Exporting Dataset
 - Data Wrangling Task 4.4 (with same subpoints as in the exercise)

# 01. Importing Libraries

import pandas as pd
import numpy as np
import os 

# 02. Importing Dataset

path = r'C:\Users\viki\Documents\Data Analytics\Immersion\Achievement 4\Instacart Basket Analysis'

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)


#checking shape of all three dfs
df_ords.shape

df_prods.shape

# 03. DATA WRANGLING IN THE EXERCISE

# 03.a - Dropping Columns

#dropping (visually) an unnecessary column
df_ords.drop(columns = ['eval_set'])

#overwriting orders dataframe (dropping column)
df_ords = df_ords.drop(columns = ['eval_set'])

df_ords.head()

#looking for NAs
df_ords['days_since_prior_order'].value_counts(dropna = False)

# 03.b - Renaming Columns

# renaming column
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head ()

# 03.c - Changing Variables' Datatype

# changing column "order_id" and "user_id" datatype to string
df_ords['order_id'] = df_ords['order_id'].astype('str')
df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['order_id'].dtype

df_ords['user_id'].dtype

# 03.d - Transposing Data

#checking department dataframe
df_dep.head()

#transposing df_dep
df_dep.T

#creating new df for the transposed version
df_dep_t = df_dep.T

#checking new version what it looks like
df_dep_t

#creating index
df_dep_t.reset_index()

#taking the first row of df_dep_t to be a new header
new_header = df_dep_t.iloc[0]

new_header

#removing first row
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

#set the header row as the df header
df_dep_t_new.columns = new_header 

#checking our final transposed df with new header
df_dep_t_new

# 03.e - Data Dictionary

#creating data dict
data_dict = df_dep_t_new.to_dict('index')

data_dict

# 03.f - Subsetting

df_prods.head()# checking our df

print(data_dict.get('19'))#what does 19 stand for in our data dictionary?

# creating snacks subset
df_snacks =  df_prods[df_prods['department_id']==19]

df_snacks.head()

# 03.f - Exporting Dataframes

# Exporting df at the end 
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

#  04. DATA WRANGLING TASKS

# Step 2 - Find another identifier variable in the df_ords dataframe that doesn’t need to be included in your analysis as a numeric variable and change it to a suitable format.

df_ords.head()

# changing column "user_id" datatype to string
df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

# Step 3 - Look for a variable in your df_ords dataframe with an unintuitive name and change its name without overwriting the data frame.

# renaming column - we alredy overwrote "dow" during the exercise and the others are intuitive, so we just change one column name to practice
df_ords.rename(columns = {'days_since_prior_order' : 'days_since_previous_order'}, inplace = False)

# Step 4 - Your client wants to know what the busiest hour is for placing orders. Find the frequency of the corresponding variable and share your findings.

df_ords['order_hour_of_day'].value_counts(dropna = False) # we can see that 10 in the moring is the busiest hour

# Step 5 - Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary.

print(data_dict.get('4')) # data dictionary have been created during the exercise already, value of 4 stands for "produce"

# Step 6 - The sales team in your client’s organization wants to know more about breakfast item sales. Create a subset containing only the required information.

# creating breakfast subset
df_breakfast =  df_prods[df_prods['department_id']==14]

df_breakfast.head()

# Step 7 - They’d also like to see details about customers who might be throwing dinner parties. Your task is to find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood. You’ll need to present this subset to your client.

# creating subset with required departments
df_dinnerparties =  df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

df_dinnerparties #presenting the new df

# Step 8 - It’s important that you keep track of total counts in your dataframes. How many rows does the last dataframe you created have?

# we can see in the above query that it has 7650 rows

# Step 9 - Someone from the data engineers team in Instacart thinks they’ve spotted something strange about the customer with a "user_id" of “1.” Extract all the information you can about this user.

# creating subset for the purchases of user 1
df_user1 =  df_ords[df_ords['user_id']=='1']

df_user1

# Step 10 - You also need to provide some details about this user’s behavior. What basic stats can you provide based on the information you have?


df_user1.describe()# ordered 11 times, always between 7 and 16 o'clock, in an average every 20 days

# Step 12/13 - Export your df_ords dataframe as “orders_wrangled.csv” in your “Prepared Data” folder. Export the df_dep_t_new dataframe as “departments_wrangled.csv” in your “Prepared Data” folder

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv')) #exporting df_ords

df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv')) #exporting df_dep_t
