==> Scripts/3. Data consistency checks.ipynb <==
# 00. Table of contents
 - Importing libraries
 - Importing Dataset
 - Looking for mixed type data and addressing them
 - Missing values and addressing them
 - Duplicates and addressing them

# 01. Importing libraries

import pandas as pd
import numpy as np
import os

# 02. Importing Dataset


path = r'C:\Users\viki\Documents\Data Analytics\Immersion\Achievement 4\Instacart Basket Analysis'


df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = 0)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# 03. Test Exercise

## 03.a - Mixed type data

# creating a new df
df_test = pd.DataFrame()

#adding new column w/ mixed values
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# check for mixed types
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# change datatype to string
df_test['mix'] = df_test['mix'].astype('str')

df_test['mix'].dtype

## 03.b - Missing values

df_prods.isnull().sum()

#creating a subset of rows with missing values
df_nan = df_prods[df_prods['product_name'].isnull()== True]

df_nan

#checking number of rows
df_prods.shape

# removing rows with NAs
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

## 03.c - Duplicates

#looking for duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

#checking No. of rows
df_prods_clean.shape

#dropping duplicates by creating new df
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

#checking No.of rows in new df
df_prods_clean_no_dups.shape

# 04. Task

## Step 2 
### Run the df.describe() function on your df_prods dataframe. Using your new knowledge about how to interpret the output of this function, share in a markdown cell whether anything about the data looks off or should be investigated further. 

df_prods.describe().round(2)

<font color=blue>The maximum price of 99999 seems unreasonal, this could be further investigated</font> 

## Step 3
### Check for mixed-type data in your df_ords dataframe.

for col in df_ords.columns.tolist():
  weird2 = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird2]) > 0:
    print (col)

<font color=blue>The command didnt find any mixed type data. QUESTION: if I used earlier in the exercise the variable 'weird' than here it is better to use a new name for the variable (weird2), right? does it have a significance ot it would be okay to use here "weird" again as I am running the two cells separately anyway?</font> 

## Step 4 
### If you find mixed-type data, fix it. The column in question should contain observations of a single data type.

<font color=blue>There were none</font> 

## Step 5 
### Run a check for missing values in your df_ords dataframe. In a markdown cell, report your findings and propose an explanation for any missing values you find.

#looking for missing values
df_ords.isnull().sum() 

#checking No. of rows in the df
df_ords.shape

#checking frequency on the column with missing values
df_ords['days_since_prior_order'].value_counts(dropna = False)

<font color=blue>Approximately 6% of the values are missing in the "days since prior order" column.The reason for the missing value could be that that order was the very first of the customer so the value should be 0. But we see in the fequency table that we have 0 values in our column, so that is probably not the real reason why it is missing. It could be just randomly missing by forgetting to record the value during the data collection</font> 

## Step 6 
### Address the missing values using an appropriate method. In a markdown cell, explain why you used your method of choice.

#looking for mean
df_ords['days_since_prior_order'].mean()

#imputing mean (11) for the missing values
df_ords_clean = df_ords.copy()
df_ords_clean['days_since_prior_order'] = df_ords['days_since_prior_order'].fillna(11)
df_ords_clean

#checking if NaNs got imputed in the new df
df_ords_clean['days_since_prior_order'].value_counts(dropna = False)

df_ords_clean

<font color=blue>As there were no extreme outliers, I imputed the missing values with the mean
QUESTION: my script has an issue: it overwrites also the NaNs in the original df and this is not what I wanted.I wanted to create this clean version and leave the original unchanged. What should I do differently to achieve this?</font> 

## Step 7 
### Run a check for duplicate values in your df_ords data. In a markdown cell, report your findings and propose an explanation for any duplicate values you find.


#finding duplicates
df_dups2 = df_ords_clean[df_ords_clean.duplicated()]

#viewing duplicates
df_dups2

<font color=blue>There are no duplicates</font> 

## Step 8  
### Address the duplicates using an appropriate method. In a markdown cell, explain why you used your method of choice.


<font color=blue>There are no duplicates</font>

## Step 9 
### Export your final, cleaned df_prods and df_ords data as “.csv” files in your “Prepared Data” folder and give them appropriate, succinct names.

df_ords_clean.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_cleaned_final.csv')) #exporting df_ords_clean

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_cleaned_final.csv')) #exporting df_ords_clean_no_dups

df_ords_clean.shape


df_prods_clean_no_dups.shape
