==> Scripts/4. Combining and Exporting Data_Part1.ipynb <==
# 00. Table of contents
 - Importing libraries
 - Importing Dataset
 - Merging Dataset :orders + order_products_prior

# 01. Import Libraries

import pandas as pd
import numpy as np
import os

# 02. Importing Datasets

path = r'C:\Users\viki\Documents\Data Analytics\Immersion\Achievement 4\Instacart Basket Analysis'

df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'order_products__prior.csv'), index_col = False)

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_cleaned_final.csv'), index_col = 0)


# 03. Merge Datasets

### consistency checks for order_products__prior df

#looking for mixed types
for col in df_ords_prior.columns.tolist():
  weird = (df_ords_prior[[col]].applymap(type) != df_ords_prior[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords_prior[weird2]) > 0:
    print (col)

#looking for missing values
df_ords_prior.isnull().sum() 

#looking for duplicates
df_dups = df_ords_prior[df_ords_prior.duplicated()]

#viewing duplicates
df_dups

## Step 1 - If you haven’t merged your prepared Instacart orders data with the new orders_products_prior dataframe per the instructions in this Exercise, do so now.


#checking our dataframe
df_ords_prior.head()

#checking our other dataframe 
df_ords.head()

df_ords_prior.shape

df_ords.shape

# merging our two dataframes
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

# checking our merged dataframe
df_merged_large.head()

df_merged_large['_merge'].value_counts()

df_merged_large.shape

## Step 2 - Export the merged file in pickle format as “orders_products_combined.pkl”.

#exporting merged df into pickles format
df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
