==> Scripts/4. Combining and Exporting Data_Part2.ipynb <==
# 00. Table of contents
 - Importing libraries
 - Importing Dataset
 - Merging Datasets :orders_products_combined + products

# 01. Importing libraries

# Import libraries

import pandas as pd
import numpy as np
import os

# 02. Importing Datasets


## Step 3 - In a new notebook, import the orders_products_combined dataframe from the pickle file you just saved.

path = r'C:\Users\viki\Documents\Data Analytics\Immersion\Achievement 4\Instacart Basket Analysis'

df_orders_products_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

# 03. Merging Dataframes

## Step 4 - Check the shape of the imported dataframe 

#checking shape of imported df
df_orders_products_combined.shape

## Step 5 - Determine a suitable way to combine the orders_products_combined dataframe with your products data set. Make sure you’re using your wrangled, cleaned, and deduped products data set stored in your “Prepared Data” folder from the previous Exercise’s task.

#importing the wrangled products df
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_cleaned_final.csv'), index_col = 0)

#viewing df prods
df_prods.head()

#viewing df_orders_products_combined 
df_orders_products_combined.head()

#dropping merge flag from above df
df_orders_products_combined= df_orders_products_combined.drop(columns=['_merge'])

#merging the previous two dfs
df_orders_products_merged = df_orders_products_combined.merge(df_prods, on = 'product_id',indicator= True)

#viewing final merged df
df_orders_products_merged.head()

#viewing shape of merged df
df_orders_products_merged.shape

## Step 6. Confirm the results of the merge using the merge flag.

df_orders_products_merged['_merge'].value_counts()

## Step 7. - Export the newly created dataframe as orders_products_merged in a suitable format (taking into consideration the size)

# dropping _merge column to avoid problems when merging again
df_orders_products_merged.drop(columns = ['_merge'])

df_orders_products_merged.head()

#exporting final merged df in pickle format
df_orders_products_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
