==> Scripts/5. Deriving New Variables.ipynb <==
# 00. Table of contents
 - Importing libraries
 - Importing Dataset
 - Creating price labels (with IF function) for a subset
 - Creating price labels (with loc function) for the entire merged dataframe
 - Creating new column for flagging how busy a day is (2 versions: busiest day and busiest days)
 - Creating new column for labeling period of day how busy is based on number of orders

# 01. Importing libraries

# Import libraries

import pandas as pd
import numpy as np
import os

# 02. Importing dataset

path = r'C:\Users\viki\Documents\Data Analytics\Immersion\Achievement 4\Instacart Basket Analysis'

df_ords_prods_merge= pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))

df_ords_prods_merge.shape

# 03. Creating a subset to avoid memory issues

#creating subset of the first 1 million rows
df = df_ords_prods_merge[:1000000]

# 04. Creating price labels with IF statements


# defining new variable
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# creating new column for price range
df['price_range'] = df.apply(price_label, axis=1)

#checking frequency in our new column
df['price_range'].value_counts()

#looking for the most expensive product
df['prices'].max()

# 05. creating price labels with loc for the entire df

df_ords_prods_merge.loc[df_ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods_merge.loc[(df_ords_prods_merge['prices'] <= 15) & (df_ords_prods_merge['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df_ords_prods_merge.loc[df_ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

#checing frequency on the newly created column
df_ords_prods_merge['price_range_loc'].value_counts()

# 06. analysing how busy each dow is

# checking frequency of dow
df_ords_prods_merge['orders_day_of_week'].value_counts(dropna=False)

#creating a list to categorise how busy a day is
result = []

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

#creating a new column with the result list
df_ords_prods_merge['busiest_day']= result

#checking frequency on the new column
df_ords_prods_merge['busiest_day'].value_counts(dropna=False)

# 07. Task
## Step 2
### Suppose your clients have changed their minds about the labels you created in your “busiest_day” column. Now, they want “Busiest day” to become “Busiest days” (plural). This label should correspond with the two busiest days of the week as opposed to the single busiest day. At the same time, they’d also like to know the two slowest days. Create a new column for this using a suitable method.

#creating a list to categorise busyness
result2 = []

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value ==0 or value== 1:
    result2.append("Busiest days")
  elif value  ==3 or value== 4:
    result2.append("Least busy days")
  else:
    result2.append("Regularly busy")

#creating a new column with the result2 list
df_ords_prods_merge['busiest_days']= result2

## Step 3
### Check the values of this new column for accuracy. Note any observations in markdown format.

#checking frequency on the new column regarding business - output is in percentage
df_ords_prods_merge['busiest_days'].value_counts(dropna=False, normalize=True, ascending= True)

#checking frequency on the "old"column regarding business - output is in percentage
df_ords_prods_merge['busiest_day'].value_counts(dropna=False, normalize=True, ascending= True)

<font color=blue>Not sure what exactly is meant by checking the accuracy. I compared how significant difference there is within the result in case we change how many days we consider as busy / regular / least busy. We can see that the weighting among the categories have changed significantly: regulary busy label is being not that outstanding anymore </font> 

## Step 4
### When too many users make Instacart orders at the same time, the app freezes. The senior technical officer at Instacart wants you to identify the busiest hours of the day. Rather than by hour, they want periods of time labeled “Most orders,” “Average orders,” and “Fewest orders.” Create a new column containing these labels called “busiest_period_of_day.”

#checking what order hour column exactly called
df_ords_prods_merge.columns

#checking frequency for hours
df_ords_prods_merge['order_hour_of_day'].value_counts(dropna=False)

#creating list for busiest hour categories
result3 = []

for value in df_ords_prods_merge["order_hour_of_day"]:
  if value in [10,11,14,15,13,12,16,9]:
    result3.append("Most orders")
  elif value in [3,4,2,5,1,0,6,23]:
    result3.append("Fewest orders")
  else:
    result3.append("Average orders")

#creating a new column with the result3 list
df_ords_prods_merge['busiest_period_of_day']= result3

## Step 5
### Print the frequency for this new column.

#printing frequency
df_ords_prods_merge['busiest_period_of_day'].value_counts(dropna=False)

## Step 7
### Export your dataframe as a pickle file (since you added new columns) and store it correctly in your “Prepared Data” folder.

#checking all columns in the altered df
df_ords_prods_merge.columns

df_ords_prods_merge.shape

#exporting final df with new columns in pickle format
df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged_with_new_columns.pkl'))
