==> Scripts/7. Data Visualisation_Part2.ipynb <==
# 00. Table of contents
 - Importing libraries
 - Importing Dataset
 - Data Wrangling for new df: customers
 - Data consistency checks for new df: customers
 - Exporting dfs as a preparation for merge in a separate notebook

# 01. Importing libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# 02. Importing dataset

path = r'C:\Users\viki\Documents\Data Analytics\Immersion\Achievement 4\Instacart Basket Analysis'

df_customers= pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

ords_prods_merge= pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_4_9.pkl'))

# 03. Data Wrangling (Step 4)

### Exploring the new df

df_customers.head()

df_customers.describe()

df_customers.shape

df_customers.dtypes

### Checking frequencies

df_customers['Gender'].value_counts()

df_customers['STATE'].value_counts().sort_index()

df_customers['date_joined'].value_counts().sort_index()

df_customers['n_dependants'].value_counts().sort_index()

df_customers['fam_status'].value_counts().sort_index()

### Let's do some wrangling

# changing column names to be more intuitive
df_customers.rename(columns={'Surnam':'Surname'}, inplace= True)
df_customers.rename(columns={'STATE':'State'}, inplace= True)
df_customers.rename(columns={'income':'Income'}, inplace= True)

#checking renaming results
df_customers.head()

# changing user_id datatype to string
df_customers['user_id']=df_customers['user_id'].astype('str')

# checking the result of changing datatype
df_customers['user_id'].dtype

# 04. Data quality and consistency (Step 5)

# looking for mixed type data

for col in df_customers.columns.tolist():
  weird = (df_customers[[col]].applymap(type) != df_customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_customers[weird]) > 0:
    print (col)

# changing dtype for column "first name"
df_customers['First Name']=df_customers['First Name'].astype('str')

# checking for missing values
df_customers.isnull().sum()

<font color=blue>There were no missing values in our df</font> 


# looking for duplicates by creating a subset
df_dups= df_customers[df_customers.duplicated()]

df_dups.head()

<font color=blue>There were no duplicates in our df</font> 

# 06. Merging Data (Step 6)

# checking our other df
ords_prods_merge.head()

#checking columns in our other df 
ords_prods_merge.columns

ords_prods_merge.shape

df_customers.shape

# we want to do a merge on user_id, double checking dtype in the ords_prods_merge df
ords_prods_merge['user_id'].dtype

# changing datatype to string
ords_prods_merge['user_id'] = ords_prods_merge['user_id'].astype('str')

# checking datatype after changing
ords_prods_merge['user_id'].dtype

# exporting altered datasets - will try to merge them in a new notebook as here it failed due to memory issues no matter what I tried
ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged_for_final_merge.pkl'))
df_customers.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'customers_for_final_merge.pkl'))
