==> Scripts/7. Data Visualisation_Part3_merging.ipynb <==
# 00. Table of contents
 - Importing libraries
 - Importing Dataset
 - Merging dfs: orders_products_merged + customers

# Importing libraries

import pandas as pd
import numpy as np
import os

# Importing datasets


path = r'C:\Users\viki\Documents\Data Analytics\Immersion\Achievement 4\Instacart Basket Analysis'

ords_prods_merge= pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_for_final_merge.pkl'))
df_customers= pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'customers_for_final_merge.pkl'))

ords_prods_merge.shape


df_customers.shape

#merging the two dfs on user_id
ords_prods_cust_merged = ords_prods_merge.merge(df_customers, on = 'user_id')

ords_prods_cust_merged.shape

ords_prods_cust_merged.head()

# Exporting merged dataset

ords_prods_cust_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_cust_merged.pkl'))
